requirements.txt
PySimpleGUI==4.70.1
pyodbc
python-dotenv
